+++ include/tc/core/polyhedral/cuda/mapped_scop.h
@@ -23,7 +23,7 @@
 
 #include "tc/core/cuda/cuda_mapping_options.h"
 #include "tc/core/polyhedral/cuda/cuda_mapping_types.h"
-#include "tc/core/polyhedral/memory_promotion_heuristic.h"
+#include "tc/core/polyhedral/cuda/memory_promotion_heuristic.h"
 #include "tc/core/polyhedral/scop.h"
 #include "tc/core/utils/dlpack.h"
 #include "tc/external/isl.h"
