+++ src/core/cpu/cpu_tc_executor.cc
@@ -17,7 +17,7 @@
 
 #include "tc/core/cpu/cpu_mapping_options.h"
 #include "tc/core/halide_utils.h"
-#include "tc/core/polyhedral/mapped_scop.h"
+#include "tc/core/polyhedral/cuda/mapped_scop.h"
 #include "tc/core/tc2halide.h"
 #include "tc/core/utils/dlpack.h"
