+++ src/core/polyhedral/reduction_matcher.cc
@@ -17,7 +17,7 @@
 
 #include <unordered_set>
 
-#include "tc/core/polyhedral/mapped_scop.h"
+#include "tc/core/polyhedral/cuda/mapped_scop.h"
 #include "tc/core/polyhedral/schedule_tree.h"
 #include "tc/core/polyhedral/scop.h"
 #include "tc/external/isl.h"
