+++ test/test_mapper_memory_promotion.cc
@@ -17,11 +17,11 @@
 #include <glog/logging.h>
 #include <gtest/gtest.h>
 
-#include "tc/core/polyhedral/codegen_cuda.h"
+#include "tc/core/polyhedral/cuda/codegen.h"
 #include "tc/core/polyhedral/exceptions.h"
-#include "tc/core/polyhedral/mapped_scop.h"
+#include "tc/core/polyhedral/cuda/mapped_scop.h"
 #include "tc/core/polyhedral/memory_promotion.h"
-#include "tc/core/polyhedral/memory_promotion_heuristic.h"
+#include "tc/core/polyhedral/cuda/memory_promotion_heuristic.h"
 #include "tc/core/polyhedral/scop.h"
 
 using namespace std;
